# Add WB_PROTECTED to mutexes