@ko1 2021-09-17
```C
#define mutex_mark ((void(*)(void*))0)
```

👍